migrated changes to TestConjectureModule
